third_party/WebKit/LayoutTests/hittesting/culled-inline-crash.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
* {
  line-height: 12px;
}
li {
  position: relative;
  padding-top: 0.3px;
}
</style>
<div id="container">
<ul>
<li><a href="#">HoverHoverHoverHover</a></li>
<li><a href="#">HoverHoverHoverHover</a></li>
<li><a href="#">HoverHoverHoverHover</a></li>
<li><a href="#">HoverHoverHoverHover</a></li>
<li><a href="#">HoverHoverHoverHover</a></li>
<li><a href="#">HoverHoverHoverHover</a></li>
<li><a href="#">HoverHoverHoverHover</a></li>
<li><a href="#">HoverHoverHoverHover</a></li>
</ul>
</div>
<div id="log"></div>
<script>
test(function(t)
{
  assert_equals(document.elementFromPoint(139, 49).tagName, 'A');
}, "This test passes if it does not crash.");
</script>
